Move `buffered_scheduler` out of local
It's got nothing to do with "local", it is reusable just fine as-is in
distributed scenarios.
